requirements.txt
openai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
Faker>=40.0.0
rstr>=3.2.2
imageio-ffmpeg>=0.6.0
playwright>=1.52.0
rembg[cpu]>=2.0.74; python_version < "3.14"
SQLAlchemy>=2.0.0
pytest>=8.3.0
